Create protocols module containing log_height sync protocol
